type mistake

+++ Geometric/MathConstants.h
@@ -3,6 +3,6 @@
 namespace bad
 {
 	static constexpr float EPSILON = 0.0001f;
-	static constexpr float PI = 3.14159265358979323846f;
-	static constexpr float TAU = 6.28318530717958647692f;
+	static constexpr double PI = 3.14159265358979323846f;
+	static constexpr double TAU = 6.28318530717958647692f;
 }
